[Simple Delay] My first custom effect for the NTS-1
This implements a very basic mono delay, where the first frame sample is
delayed and mirrored to both output samples with a (settable) delay time
and feedback amount.

This is more like a controllable looper in function because there is no
additional diffusion applied to "smooth out" the sound over time, or
stereo handling. When the feedback is set to the max level, the delay
buffer just keeps playing back at the input amount.